finish exercise 1.8 and 1.9
